#17 Add SET_PEER_ID action
