+++ app/src/store/reducer.ts
@@ -1,5 +1,6 @@
 import {Reducer} from 'redux'
-import {SeatAction, SeatState} from "./types";
+import {SeatState} from "./types";
+import {SeatAction} from "./actions";
 
 const initialState: SeatState = {
     init: false
@@ -7,15 +8,11 @@ const initialState: SeatState = {
 
 export const seatReducer: Reducer<SeatState> = (state = initialState, action) => {
     switch (action.type) {
+        case SeatAction.SET_PEER_ID: {
+            return {...state, peerId: action.payload}
+        }
         case SeatAction.CURRENT_STATE_RESPONSE: {
-            // update based on timestamp
-            console.log('[REDUCER] CURRENT STATE RESPONSE')
-            // console.log(action.payload)
-            // console.log(Object.assign({}, state, action.payload, {init: true}))
-            // return Object.assign({}, state, action.payload, {init: true})
-            return {
-                ...state, init: true
-            }
+            return {...state, ...action.payload, init: true}
         }
         default: {
             return state
